wip: Improve code quality
