+++ demo/components/ConcurrentLazyLoading/index.tsx
@@ -1,0 +1,15 @@
+import {Suspense, lazy} from 'react';
+
+const LazyComponent = lazy(() => import('./components/LazyComponent'));
+
+function ConcurrentLazyLoading() {
+    return (
+        <>
+            <Suspense fallback={<h1>Component Loading</h1>}>
+                <LazyComponent />
+            </Suspense>
+        </>
+    );
+}
+
+export default ConcurrentLazyLoading;
